requirements.txt
aiofiles==23.1.0
alembic==1.11.1
amqp==5.1.1
anyio==3.6.2
appnope==0.1.3
asgiref==3.7.2
asttokens==2.2.1 
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==4.0.1
billiard==3.6.4.0
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
comm==0.1.2
cryptography==41.0.1
databases==0.7.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
Django==4.2.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
executing==1.2.0
fastapi==0.95.1
fastapi-pagination==0.12.4
filelock==3.12.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
identify==2.5.24
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
iniconfig==2.0.0
ipykernel==6.21.1
ipython==8.10.0
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.0.2
jupyter_core==5.2.0
kombu==5.2.4
labelImg==1.8.6
lxml==4.9.2
Mako==1.2.4
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mysql-connector-python==8.0.33
# nest-asyncio==1.5.6
nodeenv==1.7.0
numpy==1.24.2
oauthlib==3.2.2
opencv-python==4.7.0.72
packaging==23.0
parso==0.8.3
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.5.0
pluggy==1.0.0
pre-commit==3.3.1
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.7
Pygments==2.14.0
PyJWT==2.7.0
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.0
redis==4.5.4
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
SQLAlchemy==1.4.48
sqlparse==0.4.4
stack-data==0.6.2
starlette==0.26.1
tomli==2.0.1
tornado==6.2
traitlets==5.9.0
typing_extensions==4.5.0
unicorn==2.0.1.post1
urllib3==2.0.2
uvicorn==0.22.0
vine==5.0.0
virtualenv==20.23.0
wcwidth==0.2.6
zipp==3.13.0
django-cors-headers
django_rest_passwordreset